happy path tests
